requirements.txt
django==1.8
djangorestframework
django-cors-headers
ipython
requests
